Deployment/requirements.txt
streamlit 
pandas 
seaborn
numpy
matplotlib
tensorflow
plotly
Pillow
numpy
scipy
scikit-learn==1.2.1